requirements.txt
whisper
python-dotenv
pydub
wave
pyaudio